refactor(transformations): transformation editor barrelling and readme
