+++ apps/app/src/routes/(config)/transformations/editor/index.ts
@@ -1,0 +1,2 @@
+export { default as Editor } from './Editor.svelte';
+export { default as Runs } from './Runs.svelte';
