+++ apps/app/src/routes/(config)/transformations/new/+page.svelte
@@ -6,7 +6,7 @@
 	import { generateDefaultTransformation } from '$lib/services/db';
 	import { toast } from '$lib/toast';
 	import { createMutation } from '@tanstack/svelte-query';
-	import RenderTransformation from '../-components/RenderTransformation.svelte';
+	import { Editor } from '../editor';
 
 	const createTransformation = createMutation(
 		rpc.transformations.mutations.createTransformation.options,
@@ -23,7 +23,7 @@
 		</Card.Description>
 	</Card.Header>
 	<Card.Content class="space-y-6">
-		<RenderTransformation
+		<Editor
 			{transformation}
 			setTransformation={(newTransformation) => {
 				transformation = newTransformation;
